backend/requirements.txt
fastapi
uvicorn
httpx
openai
google-genai
python-multipart

